Split up functional test files
